feat(lint): add useForComponent
