+++ .changeset/twelve-plants-marry.md
@@ -2,4 +2,4 @@
 "@biomejs/biome": minor
 ---
 
-Add the new rule `usePrefer`, which enforce using Solid's `<For />` component for mapping an array to JSX elements.
+Add the new rule `useForComponent`, which enforce using Solid's `<For />` component for mapping an array to JSX elements.
