+++ crates/biome_cli/src/execute/migrate/eslint_any_rule_to_biome.rs
@@ -2123,7 +2123,7 @@ pub(crate) fn migrate_eslint_any_rule(
             let group = rules.nursery.get_or_insert_with(Default::default);
             let rule = group
                 .unwrap_group_as_mut()
-                .use_prefer_for
+                .use_for_component
                 .get_or_insert(Default::default());
             rule.set_level(rule.level().max(rule_severity.into()));
         }
