+++ crates/biome_js_analyze/tests/specs/nursery/useForComponent/invalid.tsx
