+++ crates/biome_js_analyze/tests/specs/nursery/useForComponent/valid.tsx
